+++ infra/Variable.hpp
@@ -75,7 +75,7 @@ class Variable {
   double GetValue(const T& object) const;
 
   template<class A, class B>
-  double GetValue(const A& a, int a_id, const B& b, int b_id) const;
+  double GetValue(const A& a, size_t a_id, const B& b, size_t b_id) const;
 
   void SetName(std::string name) { name_ = std::move(name); }
   void SetSize(short size) { size_ = size; }
@@ -87,7 +87,7 @@ class Variable {
   std::string name_;
   std::vector<Field> fields_{};
   std::set<std::string> branch_names_{};
-  std::set<short> branch_ids_{};
+  std::set<size_t> branch_ids_{};
   mutable std::vector<double> vars_{}; //!
   std::function<double(std::vector<double>&)> lambda_{[](std::vector<double>& var) { return var.at(0); }};//!
 
@@ -109,7 +109,7 @@ double Variable::GetValue(const T& object) const {
 }
 
 template<class A, class B>
-double Variable::GetValue(const A& a, int a_id, const B& b, int b_id) const {
+double Variable::GetValue(const A& a, size_t a_id, const B& b, size_t b_id) const {
   //    assert(branch_ids_.size() == 2);
   vars_.clear();
   for (const auto& field : fields_) {
